requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-filter==22.1
django-groupadmin-users==0.3.4
django-htmx==1.13.0
django-registration-redux==2.12
django-settings-export==1.2.1
django-tables2==2.5.2
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
Faker==17.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.3
